refactor: Product에서 Cart로 변경

+++ app/src/main/java/woowacourse/shopping/data/local/cart/CartDao.kt
@@ -48,7 +48,7 @@ interface CartDao {
     fun getTotalQuantity(): Int
 
     @Query("SELECT * FROM cart WHERE id = :id LIMIT 1")
-    fun findById(id: Long): CartEntity?
+    fun findById(id: Int): CartEntity?
 
     @Update
     fun update(cartEntity: CartEntity)
